Refactor: Introduce new main menu and improve Hater Mode
This commit introduces a new main menu system with a collapsible navigation rail and an expandable drawer for more options. It also significantly refactors the Hater Mode ("Magic 8-Ball") feature, including a new physics-based animation for the die and a revamped UI.

**Key Changes:**

**Main Menu & Navigation:**
-   **`MainLayout.kt`:** (New file) Organizes the main screen structure, incorporating the new menu system.
-   **`ExpressiveNavigationRail.kt`:** (New file) Implements a collapsible navigation rail for quick actions.
-   **`MenuDrawerContent.kt`:** (Existing, modified) Content for the new expandable menu drawer.
-   **`UiModel.kt` (`CueDetatState`, `MainScreenEvent`):**
    -   Added `isMenuVisible` and `isExpandedMenuVisible` to `CueDetatState` to manage menu visibility.
    -   Added `ToggleMenu` and `ToggleExpandedMenu` events.
    -   Removed `MenuClosed` event.
-   **`ToggleReducer.kt`:** Handles the new `ToggleMenu` and `ToggleExpandedMenu` events.
-   **`ProtractorScreen.kt`:**
    -   Integrates `MainLayout` to display UI chrome (menu, buttons, sliders) around the `ProtractorOverlay`.
    -   Removed direct gesture handling from `ProtractorOverlay`; it's now in `ProtractorOverlay.kt`.

**Hater Mode (Magic 8-Ball) Refactor:**
-   **`HaterViewModel.kt`:**
    -   Rewritten to manage a physics-based die animation using `kphysics`.
    -   Introduced `TriangleState` (IDLE, SUBMERGING, EMERGING, SETTLING) to manage animation.
    -   Manages a list of answers, ensuring they are not repeated too quickly.
    -   Handles shake gestures via `SensorDataManager`.
-   **`HaterState.kt`:** (New file, replaces parts of old `HaterState.kt` logic)
    -   Composable function that implements the Hater Mode UI.
    -   Manages the `kphysics` world, die body, and screen boundaries.
    -   Handles physics simulation and rendering of the die and answer text.
    -   Includes logic for resizing the die and text layout based on the answer content.
    -   Uses `SensorDataManager` for shake detection.
-   **`SensorDataManager.kt`:** (New file) Manages accelerometer sensor data to detect shake gestures.
-   **`HaterScreen.kt`:**
    -   Updated to use the new `HaterViewModel` and `HaterState`.
    -   Renders the 8-ball appearance with a viewing portal and animated die.
    -   Allows users to "push" the die by dragging on the screen.
-   **`OverlayState.kt`:** Removed. Its contents related to Hater Mode are now in the new `HaterViewModel` and `HaterState`.
-   **`MainViewModel.kt`:** Imports the new `HaterViewModel`.

**Other Notable Changes:**
-   **`VisionRepository.kt`:**
    -   Fixed a potential memory leak/crash by correctly handling `Mat` objects during image rotation. Ensures original `Mat` is released if a new rotated `Mat` is created.
-   **`ProtractorOverlay.kt`:**
    -   Gesture detection logic moved here from `ProtractorScreen.kt` using `detectManualGestures`.
-   **`QuickAlignViewModel.kt` (renamed from `QuickAlignAnalyzer.kt`):**
    -   Now a `ViewModel` (`QuickAlignViewModel`).
    -   Manages the state for the quick alignment process (pan, zoom, rotation).
    -   Introduced `QuickAlignStep` (SELECT_SIZE, ALIGN_TABLE).
    -   Fixed a race condition by resetting state *after* emitting alignment results.
-   **`MainActivity.kt`:**
    -   If camera permission is denied, it now shows a Toast message and finishes the activity.
-   **`AdvancedOptionsReducer.kt`:** Type cast `action.value` to `Int` for `houghThreshold`.
-   **`build.gradle.kts` & `libs.versions.toml`:** Added `sceneform-base` dependency (though its usage is not apparent in these diffs, it might be for future Hater Mode enhancements or other features).

+++ app/src/main/java/com/hereliesaz/cuedetat/domain/reducers/AdvancedOptionsReducer.kt
@@ -54,7 +54,7 @@ internal fun reduceAdvancedOptionsAction(
 
         is MainScreenEvent.UpdateHoughP1 -> state.copy(houghP1 = action.value)
         is MainScreenEvent.UpdateHoughP2 -> state.copy(houghP2 = action.value)
-        is MainScreenEvent.UpdateHoughThreshold -> state.copy(houghThreshold = action.value)
+        is MainScreenEvent.UpdateHoughThreshold -> state.copy(houghThreshold = action.value.toInt())
         is MainScreenEvent.UpdateCannyT1 -> state.copy(cannyThreshold1 = action.value)
         is MainScreenEvent.UpdateCannyT2 -> state.copy(cannyThreshold2 = action.value)
         else -> state
